% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidian.R
\name{euclidian}
\alias{euclidian}
\title{Compute GCD using the Euclidian algorithm}
\usage{
euclidian(x, y)
}
\arguments{
\item{x}{First input scalar numeric value.}

\item{y}{Second input scalar numeric value.}
}
\value{
The greatest common divisor between x and y, a numeric value.
}
\description{
´euclidian()´ calculates the greatest common devisor (GCD),
of two scalar values.
}
\details{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
\examples{
euclidian(252,105)
}
