\name{distanceFunctions-package}
\alias{distanceFunctions-package}
\alias{distanceFunctions}
\docType{package}
\title{
\packageTitle{distanceFunctions}
}
\description{
\packageDescription{distanceFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{distanceFunctions}
\packageIndices{distanceFunctions}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{distanceFunctions}

Maintainer: \packageMaintainer{distanceFunctions}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
