\name{distanceFunctions-package}
\alias{distanceFunctions-package}
\alias{distanceFunctions}
\docType{package}
\title{
\packageTitle{distanceFunctions}
}
\description{This package makes two functions available that allow the user to
implement the Euclidean algorithm to find the greatest common divisor between
two integers, and the Dijkstra algorithm to find the shortest path between a
starting node and all other nodes in a graph.
\packageDescription{distanceFunctions}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{distanceFunctions}
\packageIndices{distanceFunctions}
}

\author{
\packageAuthor{distanceFunctions}

Maintainer: \packageMaintainer{distanceFunctions}
}
\references{
\url{https://en.wikipedia.org/wiki/Euclidean_algorithm}
\url{https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm}
}